Reorganize commands and handle datafile
Now we have three top-level commands:

  * deploy-broker
  * join
  * info
