--- pkg/subctl/cmd/create.go
@@ -1,13 +1,0 @@
-package cmd
-
-import (
-	"github.com/spf13/cobra"
-)
-
-func init() {
-	rootCmd.AddCommand(createCmd)
-}
-
-var createCmd = &cobra.Command{
-	Use: "create",
-}
